Rename package model to api since it's a more accurate name now that we only use these classes to define the api
