+++ services/basket/src/main/java/org/cryptimeleon/incentive/services/basket/BasketController.java
@@ -2,11 +2,11 @@ package org.cryptimeleon.incentive.services.basket;
 
 import lombok.extern.slf4j.Slf4j;
 import org.cryptimeleon.incentive.services.basket.exceptions.*;
-import org.cryptimeleon.incentive.services.basket.model.Basket;
-import org.cryptimeleon.incentive.services.basket.model.Item;
-import org.cryptimeleon.incentive.services.basket.model.RewardItem;
-import org.cryptimeleon.incentive.services.basket.model.requests.PutItemRequest;
-import org.cryptimeleon.incentive.services.basket.model.requests.RedeemBasketRequest;
+import org.cryptimeleon.incentive.services.basket.api.Basket;
+import org.cryptimeleon.incentive.services.basket.api.Item;
+import org.cryptimeleon.incentive.services.basket.api.RewardItem;
+import org.cryptimeleon.incentive.services.basket.api.PutItemRequest;
+import org.cryptimeleon.incentive.services.basket.api.RedeemBasketRequest;
 import org.cryptimeleon.incentive.services.basket.storage.ItemEntity;
 import org.cryptimeleon.incentive.services.basket.storage.RewardItemEntity;
 import org.springframework.beans.factory.annotation.Value;
