+++ services/basket/src/main/java/org/cryptimeleon/incentive/services/basket/api/Item.java
@@ -1,4 +1,4 @@
-package org.cryptimeleon.incentive.services.basket.model;
+package org.cryptimeleon.incentive.services.basket.api;
 
 import io.swagger.annotations.ApiModelProperty;
 import lombok.AllArgsConstructor;
@@ -6,30 +6,23 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 import org.cryptimeleon.incentive.services.basket.storage.ItemEntity;
 
+/**
+ * Dataclass representing items that be purchased in a basket.
+ */
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
-public class BasketItem {
+public class Item {
     @ApiModelProperty(value = "${item.id}")
     String id;
     @ApiModelProperty(value = "${item.title}")
     String title;
     @ApiModelProperty(value = "${item.price}")
     long price;
-    @ApiModelProperty(value = "${item.count}")
-    int count;
 
-    public BasketItem(ItemEntity itemEntity, int count) {
+    public Item(ItemEntity itemEntity) {
         this.id = itemEntity.getId();
         this.title = itemEntity.getTitle();
         this.price = itemEntity.getPrice();
-        this.count = count;
-    }
-
-    public BasketItem(Item item, int count) {
-        this.id = item.getId();
-        this.title = item.getTitle();
-        this.price = item.getPrice();
-        this.count = count;
     }
 }
