+++ services/basket/src/test/java/org/cryptimeleon/incentive/services/basket/PayBasketTest.java
@@ -1,7 +1,7 @@
 package org.cryptimeleon.incentive.services.basket;
 
-import org.cryptimeleon.incentive.services.basket.model.BasketItem;
-import org.cryptimeleon.incentive.services.basket.model.Item;
+import org.cryptimeleon.incentive.services.basket.api.BasketItem;
+import org.cryptimeleon.incentive.services.basket.api.Item;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.TestInstance;
